requirements.txt
streamlit
langchain
langchain-openai
html2text
tiktoken
python-dotenv